+++ validator/remote/src/main/java/tech/pegasys/teku/validator/remote/FailoverValidatorApiHandler.java
@@ -178,13 +178,6 @@ public class FailoverValidatorApiHandler implements ValidatorApiChannel {
   }
 
   @Override
-  public SafeFuture<Boolean> isExecutionOptimistic() {
-    return tryRequestUntilSuccess(
-        ValidatorApiChannel::isExecutionOptimistic,
-        BeaconNodeRequestLabels.GET_SYNCING_STATUS_METHOD);
-  }
-
-  @Override
   public SafeFuture<Optional<BlockContainerAndMetaData>> createUnsignedBlock(
       final UInt64 slot,
       final BLSSignature randaoReveal,