+++ validator/remote/src/main/java/tech/pegasys/teku/validator/remote/RemoteValidatorApiChannel.java
@@ -22,11 +22,5 @@ public interface RemoteValidatorApiChannel extends ValidatorApiChannel {
 
   HttpUrl getEndpoint();
 
-  @Override
-  default SafeFuture<Boolean> isExecutionOptimistic() {
-    return getSyncingStatus()
-        .thenApply(syncingStatus -> syncingStatus.getIsOptimistic().orElse(false));
-  }
-
   SafeFuture<SyncingStatus> getSyncingStatus();
 }